# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Haraj Station API
  version: 1.0.0
  description: Generated inventory of route handlers. Request and response schemas are being added per mobile contract.
servers:
  - url: https://harajstation.com
paths:
...
  "/api/mobile/listings":
    get:
      summary: "GET /api/mobile/listings"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/mobile/me/offers":
    get:
      summary: "GET /api/mobile/me/offers"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
    post:
      summary: "POST /api/mobile/me/offers"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/mobile/stores/{slug}":
    get:
      summary: "GET /api/mobile/stores/{slug}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/push/subscribe":
    post:
      summary: "POST /api/push/subscribe"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/reports":
    post:
      summary: "POST /api/reports"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/saved-searches/{id}":
    delete:
      summary: "DELETE /api/saved-searches/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/store/verify/doc/{id}":
    get:
      summary: "GET /api/store/verify/doc/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...